requirements.txt
requests
ed25519-blake2b
pyblake2==1.1.0
gmpy2
aiohttp
websocket-client==1.4.1
websockets
jsonpickle
pydot==1.4.2
tomli==2.0.1
tomli-w
oyaml
pytest 
pytest-html
nanolib
extradict
pandas
py7zr